.develop/spec.json,node,rpc: refactor openrpc registry to node/ (out of rpc/)
The openrpc service shouldn't be installed in the
rpc library itself, since the library should only provide
the actual rpc service. Service registration, as it were,
should be handled by the layer using the rpc service,
in this case node/.
L
